% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack_optimized.R
\name{greedy_knapsack_optimized}
\alias{greedy_knapsack_optimized}
\title{greedy_knapsack_optimized}
\usage{
greedy_knapsack_optimized(x, W)
}
\arguments{
\item{x}{A data frame containing two columns: 'v' (values) and 'w' (weights) of the items.
Both 'v' and 'w' must be positive numbers.}

\item{W}{A numeric value representing the maximum allowed weight in the knapsack.}
}
\value{
A list containing:
- \code{value}: The total value obtained by the selected items.
- \code{elements}: A vector of indices indicating which items were selected for the knapsack.
}
\description{
This function solves the fractional knapsack problem using a greedy approach.
This is an optimized greedy algorithm.It improves computational speed,
especially when dealing with a large number of items,
through preallocation and more efficient memory usage.
}
