% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_brute_force_knapsack.R
\name{parallel_brute_force_knapsack}
\alias{parallel_brute_force_knapsack}
\title{parallel_brute_force_knapsack}
\usage{
parallel_brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame containing two columns: 'v' (values) and 'w' (weights) of the items.
Both 'v' and 'w' must be positive numbers.}

\item{W}{A numeric value representing the maximum allowed weight in the knapsack.}

\item{parallel}{When parallel computing is used, parallel=true}
}
\value{
A list containing:
- \code{value}: The maximum total value obtained within the weight limit.
- \code{elements}: A vector of indices indicating which items are included in the optimal solution.
}
\description{
This function solves the knapsack problem using a brute-force approach.
When the data set is large, Parallelization can improve computation speed.
}
